Deal with further cases of integer literal base-vs-suffix ambiguity
Accept forms like `0x·` as exceptions to the longest match principle.

(· is in XID_Continue but not in XID_Start, and isn't a digit.)

There's no change to the official model, where the priority rule gives the
right answer for these.

But it was wrong to think that the priority rule only disambiguates between
pretokens of equal length: in this case it prefers the shorter available
pretoken.