writeup: update pretokenisation rules to allow e-suffixes
This is aiming to match the behaviour of #131656

No longer exclude 'e' from the start of the suffix for:
 - "Float literal without exponent"
 - "Integer binary literal"
 - "Integer octal literal"
 - "Integer decimal literal"

"Integer hexadecimal literal" remains unchanged.

Split the "Float literal with exponent" pretokenisation rule into
 "Float literal with signed exponent"
 "Float literal with signless exponent"
The latter now requires a digit in the exponent.

Not requiring a digit in the exponent for the former rule ensures that
forms like 1e+f32 are still rejected.