Add tree-building to the comparable implementation
Instead of flattening rustc_parse()'s output back to a sequence of
tokens, we now construct the equivalent of rustc's TokenTree and
TokenStream, and compare both the tokens and the structure.

That means we now reject ill-formed trees.

This version performs whitespace removal and token combination after
tree construction (on the theory that all the layers that consume lexer
output require well-formed trees, but proc-macros don't require token
combination).