Stop tracking spacing information in regularised tokens
This isn't really supposed to be observable information, and I don't think it
helps catch bugs in practice.

Spacing isn't naturally observable in this layer because we're comparing
variants of "Coarse" token, and the only thing spacing is used for is
deciding which fine-grained tokens can be combined.

I'm dropping it so that I won't have to bother trying to keep track of
it when constructing token trees.