Optionally lower doc-comments to attributes
New global command-line option: --lower-doc-comments

For lex_via_rustc, this means calling desugar_doc_comments().

For the comparable implementation, it's a new pass transforming the list
of FineTokens, just before tree construction.

The extent for each of the synthetic tokens is the full extent of the
original doc-comment. That's what rustc uses internally, so this lets us
continue to include extents in what 'compare' checks.