Let the grammar see the entire input
Now using a PEG repetition expression to handle the top-level matching,
rather than describing it as a separate imperative process.

Renamed the "edition nonterminal" to the "token nonterminal".

The new TOKENS_yyyy definitions are the "tokens nonterminals".

In the reimplementation the error-reporting behaviour remains
more-or-less as it was. In particular we report processing errors even
if the tokens nonterminal didn't match the entire input (which isn't
quite what the writeup would suggest).