token_matching: omit DOUBLEQUOTE, BACKSLASH, and LF from MatchData
These tokens are treated as terminals in the writeup.

The main motivation for omitting them is to stop them appearing in `inspect`
output, where they're unhelpful.

It's also the right thing in primciple, to make what the consumed() calls in
processing can see closer to what the writeup says the "characters consumed"
notation means.