Rework the description of the top-level nonterminals
Change the name "token-kind nonterminal" to "tokenisation nonterminal".

"token nonterminal" now refers to a tokenisation nonterminal which isn't a
reserved form, rather than to the TOKEN_yyy nonterminals.

Define them explicitly using the fine-grained tokens table and a list of
reserved forms, rather than saying they are the nonterminals which appear in
the top-level choice expressions.